% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{alter_label}
\alias{alter_label}
\title{Re-ordering clusters.}
\usage{
alter_label(ori_labels, re_labels)
}
\arguments{
\item{ori_labels}{Origin labels.}

\item{re_labels}{Re-ordered labels.}
}
\value{
New labels after ordering.
}
\description{
Re-ordering clusters for heatmap.
}
