% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{BuildSCE}
\alias{BuildSCE}
\title{Build SingleCellExperiment object.}
\usage{
BuildSCE(
  counts = NULL,
  normcounts = NULL,
  logcounts = NULL,
  cpm = NULL,
  tpm = NULL,
  scaleData = NULL,
  trim = 0.1,
  sparse = TRUE,
  ...
)
}
\arguments{
\item{counts}{Matrix like or data.frame object of counts with genes by cells.}

\item{normcounts}{Analogous to counts.}

\item{logcounts}{Analogous to counts.}

\item{cpm}{Analogous to counts.}

\item{tpm}{Analogous to counts.}

\item{scaleData}{Analogous to counts.}

\item{trim}{Fraction of outlier values (on each side) to be trimmed on each side.}

\item{sparse}{logical or NULL, specifying if the result should be sparse or not.}

\item{...}{Additional arguments passed on to \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}
}
\description{
Build SingleCellExperiment object.
}
