% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{compare_expr}
\alias{compare_expr}
\title{Compare gene expression}
\usage{
compare_expr(sce, genes, trunk = NULL, branch = NULL, color = NULL,
  linestyle = NULL, datatype = "logcounts", stateid = "branch",
  trajectory = "trajectory", lw = 2, span = 0.6, ...)
}
\arguments{
\item{sce}{Expression matrix.}

\item{genes}{Use the mean of bincells for smoothing.}

\item{trunk}{Normalize the smoothed value to 0-1.}

\item{ncore}{ncores for parallel.}
}
\value{
Smoothing expression.
}
\description{
Compare gene expression.
}
