% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimension.R
\name{fdl_2d}
\alias{fdl_2d}
\title{Force-directed layout}
\usage{
fdl_2d(sce, usedEmbed = "PCA", usedDims = NULL, datatype = NULL,
  filename = NULL, k = 30, seed = 1, colour_condition = NULL,
  colour_gene = NULL, color_scale = NULL, cmap = "Paired",
  title = "Force-directed layout", alpha = 0.9, size = 1, width = 4,
  height = 3)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{usedEmbed}{Specify the reduce dimension type for construct network.}

\item{usedDims}{A vector to specify particular dimension in usedEmbed to calculate.}

\item{datatype}{Type of expression.}

\item{filename}{Filename to save - default is NULL.}

\item{k}{k nearest neighbors to construct graph.}

\item{seed}{Random seed.}

\item{colour_condition}{Label the cells with continuous color.}

\item{colour_gene}{A character of a vector of a gene expression.}

\item{color_scale}{Scatter plot color correponding to levels - default is NULL.}

\item{cmap}{Color style of cmap will be used when color_scale is NULL.}

\item{title}{Plot title.}

\item{alpha}{Transparency of points.}

\item{size}{Point size.}

\item{width}{Width of the figure.}

\item{height}{Height of the figure.}
}
\value{
\itemize{
  \item Coordinate of 2-D layout
  \item ggplot2 object
}
}
\description{
2-D visualization of data on kNN graph using force-directed layout
}
