% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{FilterLow}
\alias{FilterLow}
\title{Filtering low expression}
\usage{
FilterLow(sce, minexp = 10, mincells = 5, datatype = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object or a matrix.}

\item{minexp}{Minimum expression level.}

\item{mincells}{Minimum cells expressed minexp.}

\item{datatype}{Sepcify the data type in sce for filtering.}
}
\value{
A sce object of a matrix aftering filtering.
}
\description{
Filtering the genes expressed at least minexp in less than mincells.
}
