% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{FindHVGs}
\alias{FindHVGs}
\title{Find high variable genes}
\usage{
FindHVGs(sce, datatype = "logcounts", genes.use = "filter.low",
  thr.bio = 0, thr.tech = NULL, thr.p.value = NULL, thr.FDR = 0.1,
  thr.low = 0, thr.high = 20, show.plot = TRUE, ...)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{datatype}{Sepcify the data type in sce to find hvgs.}

\item{genes.use}{A logical vector in rowData to specify the genes for calculation.}

\item{thr.bio}{The threshold of biological component of the variance.}

\item{thr.tech}{The threshold of Technical component of the variance.}

\item{thr.p.value}{The threshold of p-values for the test against the null hypothesis that bio=0.}

\item{thr.FDR}{The threshold of adjusted p-values.}

\item{thr.low}{The low threshold of mean log-expression.}

\item{thr.high}{The high threshold of mean log-expression.}

\item{show.plot}{Whether to plot the fitting results.}

\item{...}{Additional arguments passed on to \code{\link[scran]{trendVar}}.}
}
\value{
SingleCellExperiment object with hvgs mask in rowData slot.
}
\description{
Fing high variable genes by fitting a mean-dependent trend to the gene-specific variances.
}
\references{
\url{https://f1000research.com/articles/5-2122/v2}
}
