% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{gene_smooth}
\alias{gene_smooth}
\title{Smoothing expression}
\usage{
gene_smooth(mat, bincells = 1, normalize = TRUE, ncore = 1)
}
\arguments{
\item{mat}{Expression matrix.}

\item{bincells}{Use the mean of bincells for smoothing.}

\item{normalize}{Normalize the smoothed value to 0-1.}

\item{ncore}{ncores for parallel.}
}
\value{
Smoothing expression.
}
\description{
Smoothing expression using Bayesian and Likelihood Analysis of Dynamic Linear Models.
}
