% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimension.R
\name{layout_2d}
\alias{layout_2d}
\title{Force-directed layout}
\usage{
layout_2d(sce, usedEmbed = "PCA", usedDims = NULL, datatype = NULL,
  filename = NULL, k = 35, n_trees = 50, max_iter = 2, seed = 1,
  layout_method = "FR", grid = "auto", iterations = 1000,
  repel = 2, plotstep = 0, clus_method = NULL, shape = NULL,
  colour_condition = NULL, colour_gene = NULL, colour_embed = NULL,
  color_scale = NULL, cmap = NULL, compute_layout = TRUE,
  orderExp = FALSE, title = NULL, ncol = NULL, nrow = NULL,
  alpha = 0.9, size = 1, width = 4, height = 3,
  rmelements = NULL, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{usedEmbed}{Specify the reduce dimension type for construct network. See current avaible embedding by \code{\link[SingleCellExperiment]{reducedDimNames}}.}

\item{usedDims}{A vector to specify particular dimensions in usedEmbed to calculate.}

\item{datatype}{Type of expression.}

\item{filename}{Filename to save - default is NULL.}

\item{k}{k nearest neighbors to construct graph.}

\item{n_trees}{Number of trees for largeVis.}

\item{max_iter}{max_iter for largeVis. see \code{\link[SOT]{randomProjectionTreeSearch}}}

\item{seed}{Random seed.}

\item{layout_method}{Graph layout method of FR (Fruchterman-Reingold layout) or FA2 (ForceAtlas2).}

\item{grid}{Character scalar, whether to use the faster, but less accurate grid based implementation of the algorithm. By default (“auto”), the grid-based implementation is used if the graph has more than one thousand vertices.}

\item{iterations}{Iterations number of iterations to be performed.}

\item{repel}{The repel constant: the greater the constant k the stronger the repulsion force between points.}

\item{plotstep}{Plot step.}

\item{clus_method}{Clustering method besed on graph.}

\item{colour_condition}{Label the cells with continuous color.}

\item{colour_gene}{A character of a vector of a gene expression.}

\item{colour_embed}{A character of a reduced dimentsion.}

\item{color_scale}{Scatter plot color correponding to levels - default is NULL.}

\item{cmap}{Color style of cmap will be used when color_scale is NULL.}

\item{compute_layout}{Compute fr layout or not. If compute_layout is FASLE, uese the usedEmbed as layout.}

\item{orderExp}{Order plot to highlight high expressed cells.}

\item{title}{Plot title.}

\item{ncol}{(optional) number of columns in the plot grid.}

\item{nrow}{(optional) number of rows in the plot grid.}

\item{alpha}{Transparency of points.}

\item{size}{Point size.}

\item{width}{Width of the figure.}

\item{height}{Height of the figure.}

\item{rmelements}{A vector of elememts to be removed {\link[ggpubr]{rremove}}.}

\item{...}{Additional arguments passed on to \code{\link[SOT]{umap}}}
}
\value{
\itemize{
  \item Coordinate of 2-D layout
  \item ggplot2 object
  \item igraph 
}
}
\description{
2-D visualization of data on kNN graph using force-directed layout
}
