% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimension.R
\name{plot_embed}
\alias{plot_embed}
\title{Heatmap of the reduced dimension data}
\usage{
plot_embed(
  sce,
  usedEmbed,
  k = 8,
  method = "complete",
  filename = NA,
  cells.use = NULL,
  vars.use = NULL,
  anno_col = NULL,
  color_scale = NULL,
  cmap = NULL,
  cluster_row = TRUE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{usedEmbed}{Specify the reduce dimension type for construct network. See current avaible embedding by \code{\link[SingleCellExperiment]{reducedDimNames}}.}

\item{k}{Number of clusters.}

\item{method}{Clustering method.}

\item{filename}{Figure name.}

\item{anno_col}{Characters to specify annotate conditions.}

\item{color_scale}{Customer color scale for each condition.}

\item{cmap}{Color map.}
}
\value{
Clustering result.
}
\description{
Heatmap of the reduced dimension data
}
