% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{re_order}
\alias{re_order}
\title{Re-write clustering labels.}
\usage{
re_order(labs, ord_id)
}
\arguments{
\item{labs}{Clustering labels.}

\item{ord_id}{Cluster order idx.}
}
\value{
A vector of re-write labels.
}
\description{
Re-write clustering labels.
}
