% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reduce.cluster}
\alias{reduce.cluster}
\title{Reduce redundant clusters of AP clustering}
\usage{
reduce.cluster(sce, gcl, datatype = NULL, aff = NULL,
  trim = 5.1/ncol(mat), center = TRUE, scale. = TRUE, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object or matrix.}

\item{gcl}{AP clustering result for the firt round.}

\item{datatype}{Sepcify the data type in sce for filtering.}

\item{aff}{Prior defined affinity matrix - default is NULL.}

\item{trim}{Trim the largest and lowest values of each gene to avoid outlier effect.}

\item{center}{Center the expression before performing PCA - default is FALSE.}

\item{scale.}{Scale the expression before performing PCA - default is FALSE.}

\item{...}{Additional arguments passed on to \code{\link[apcluster]{apcluster}}}
}
\value{
\itemize{
  \item Group labels of genes
  \item Basic stats of Groups
  \item Eigengenes
}
}
\description{
Merge highly related clusters into groups
}
