% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimension.R
\name{umap}
\alias{umap}
\title{Uniform Manifold Approximation and Projection (UMAP).}
\usage{
umap(mat, umap.args = list(), n_neighbors = 20, n_components = 2,
  metric = "euclidean", negative_sample_rate = 5, init = "spectral",
  min_dist = 0.3, random_state = 1, verbose = FALSE)
}
\arguments{
\item{mat}{Data with rows are samples and cols are genes.}

\item{n_neighbors}{float (optional, default 15) The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation. Larger values result in more global views of the manifold, while smaller values result in more local data being preserved. In general values should be in the range 2 to 100.}

\item{n_components}{int (optional, default 2) The dimension of the space to embed into.}

\item{metric}{string (optional, default 'euclidean') The metric to use to compute distances in high dimensional space. Valid string metrics include: euclidean, manhattan, minkowski, mahalanobis, seuclidean, cosine, correlation, hamming, jaccard, kulsinski.}

\item{negative_sample_rate}{int (optional, default 5) The number of negative edge/1-simplex samples to use per positive edge/1-simplex sample in optimizing the low dimensional embedding.}

\item{init}{string (optional, default 'spectral') How to initialize the low dimensional embedding. Options are: 'spectral', 'random', A numpy array of initial embedding positions.}

\item{min_dist}{float (optional, default 0.1) The effective minimum distance between embedded points. The value should be set relative to the spread value, which determines the scale at which embedded points will be spread out.}

\item{random_state}{int, RandomState instance or NULL, optional (default: NULL)}

\item{verbose}{bool (optional, default False) Controls verbosity of logging.}
}
\value{
Low dimension embedding
}
\description{
Dimension reduction technique that can be used for visualisation similarly to t-SNE, but also for general non-linear dimension reduction.
}
