% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{wishbone}
\alias{wishbone}
\title{Wishbone pseudotime}
\usage{
wishbone(sce, usedEmbed, start_cell, datatype = "scaledata",
  usedDims = NULL, condition = NULL, usedLevels = NULL,
  plot_genes = NULL, branch = TRUE, k = 35, num_waypoints = 150,
  seed = 1, figsize = c(20, 10))
}
\arguments{
\item{sce}{A SingleCellExperiment object or matrix.}

\item{usedEmbed}{Specify the reduce dimension type for construct network. See current avaible embedding by \code{\link[SingleCellExperiment]{reducedDimNames}}.}

\item{start_cell}{Character specified a start cell of the trajectory.}

\item{datatype}{Type of expression used for plot. See current avaible assays by \code{\link[SummarizedExperiment]{assays}}.}

\item{usedDims}{A vector to specify particular dimensions in usedEmbed to calculate.}

\item{condition}{Used with usedLevels. Specify a condition in colData(sce) for trimming cells.}

\item{usedLevels}{Used with condition. Only the cells in usedLevels are used for calculating.}

\item{plot_genes}{A vector of genes id. If plot_genes is not NULL, plot the average smoothing gene expression along the trajectories.}

\item{branch}{Whether branches exist or not.}

\item{k}{kNN graph.}

\item{num_waypoints}{Number of num_waypoints.}

\item{seed}{Random seed.}

\item{figsize}{A vector correponding to the figure width and height.}
}
\value{
sce object contain pseudotime and branches.
}
\description{
Graph based pasudotime infer. Please install python package "wishbone" before using this function.
}
