% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ADtest}
\alias{ADtest}
\title{Aderson-Darling test}
\usage{
ADtest(sce, datatype = "logcounts", condition, useLevels = NULL,
  genes.use = "find.hvgs", sample.cells = 200, adj.method = "BH",
  thr.padj = 0.001, ncore = 6)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{datatype}{Sepcify the data type in sce for filtering.}

\item{condition}{Conditions corresponding to cells.}

\item{useLevels}{Subset conditions for test - default is NULL.}

\item{genes.use}{A logical vector in rowData to specify the genes for calculation.}

\item{sample.cells}{Sampling cell number in each condition to speed up testing.}

\item{adj.method}{p-value correction method. See \code{\link[stats]{p.adjust}} for more detail.}

\item{ncore}{Number of cores used for parallel.}
}
\value{
Adjusted p-value of Anderson-Darling test.
}
\description{
Test if gene expression in different sampling time points come from a common population
}
