% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{adtest}
\alias{adtest}
\title{Aderson-Darling test}
\usage{
adtest(sce, condition, useLevels = NULL, adj.method = "BH",
  datatype = NULL, ncore = 6)
}
\arguments{
\item{sce}{SingleCellExperiment object.}

\item{condition}{Conditions corresponding to cells.}

\item{useLevels}{Subset conditions for test - default is NULL.}

\item{adj.method}{p-value correction method. See \code{\link[stats]{p.adjust}} for more detail.}

\item{datatype}{Sepcify the data type in sce for filtering.}

\item{ncore}{Number of cores used for parallel.}
}
\value{
Adjusted p-value of Anderson-Darling test.
}
\description{
Test if gene expression in different sampling time points come from a common population
}
