% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ap.cluster}
\alias{ap.cluster}
\title{Affinity propagation clustering}
\usage{
ap.cluster(sce, datatype = "logcounts", genes.use = "genes.use",
  projections = "l2", dsim = NULL, trim = 5.1/ncol(sce),
  method = "pearson", center = TRUE, scale. = TRUE,
  normalize.score = FALSE, minsize = 0, ncore = 4, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object or matrix.}

\item{datatype}{Specify the data type in sce for filtering.}

\item{genes.use}{Specify the genes to perform clustering.}

\item{projections}{Whether to calculate reconstructions and scores using the L1 norm ("l1") the L2 norm ("l2").}

\item{dsim}{Prior defined distance matrix - default is NULL.}

\item{trim}{Trim the largest and lowest values of each gene to avoid outlier effect.}

\item{method}{The method to calculate the affinity matrix.}

\item{center}{Center the expression before performing PCA - default is FALSE.}

\item{scale.}{Scale the expression before performing PCA - default is FALSE.}

\item{normalize.score}{Wheter to normalize the pc1 score.}

\item{minsize}{Minimum cluster size.}

\item{ncore}{Number of cores used for parallel.}

\item{...}{Additional arguments passed on to \code{\link[apcluster]{apcluster}}}
}
\value{
\itemize{
  \item Cluster labels of genes
  \item Basic stats of clusters
  \item Eigengenes
}
}
\description{
AP clustering on genes and calculate the eigengenes of each cluster
}
