% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ap.cluster}
\alias{ap.cluster}
\title{Affinity propagation clustering}
\usage{
ap.cluster(sce, datatype = NULL, method = "cor", power = 6, aff = NULL,
  trim = 5.1/ncol(sce), center = TRUE, scale. = TRUE, minsize = 5, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object or matrix.}

\item{datatype}{Sepcify the data type in sce for filtering.}

\item{method}{The method to calculate the affinity matrix.}

\item{power}{Power level of the paire-wised correlation.}

\item{aff}{Prior defined affinity matrix - default is NULL.}

\item{trim}{Trim the largest and lowest values of each gene to avoid outlier effect.}

\item{center}{Center the expression before performing PCA - default is FALSE.}

\item{scale.}{Scale the expression before performing PCA - default is FALSE.}

\item{...}{Additional arguments passed on to \code{\link[apcluster]{apcluster}}}
}
\value{
\itemize{
  \item Cluster labels of genes
  \item Basic stats of clusters
  \item Eigengenes
}
}
\description{
AP clustering on genes and calculate the eigengenes of each cluster
}
