% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reduce.cluster}
\alias{reduce.cluster}
\title{Reduce redundant clusters of AP clustering}
\usage{
reduce.cluster(
  sce,
  datatype = "logcounts",
  genes.use = "genes.use",
  projections = "l2",
  method = "pearson",
  sim = NULL,
  center = TRUE,
  scale. = TRUE,
  normalize.score = TRUE,
  ncore = 2,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{datatype}{Sepcify the data type in sce for filtering.}

\item{method}{The method to calculate the affinity matrix, can be pearson (Defult) or cosine.}

\item{center}{Center the expression before performing PCA - default is FALSE.}

\item{scale.}{Scale the expression before performing PCA - default is FALSE.}

\item{normalize.score}{Whether to normalize pc1 score.}

\item{ncore}{Number of cores used for parallel.}

\item{...}{Additional arguments passed on to \code{\link[apcluster]{apcluster}}}

\item{dsim}{Prior defined similarty matrix of lv1.score for clustering - default is NULL.}
}
\value{
\itemize{
  \item Group labels of genes
  \item Basic stats of Groups
  \item Eigengenes
}
}
\description{
Merge highly related clusters into groups
}
